Added "num_of_columns" to Import Mappings for validating a raw file of an import job.
Currently import job validations will stop reading a file if the number of columns in the selected raw file does not match the number of columns detected during the definition of the import mapping.

Defined methods for Notifier for emailing users when starting validations or imports of import jobs.

Currently looking at storing options